+++ docs/changelog.md
@@ -20,7 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `write_output` internal helper allows for custom file extension (used for yang schema dumps)
 - `pylint` is now driven by `ruff`
 - `bandit` is now configured by pyproject.toml | updated docker file for linter + github CI
-- `SKIP` updated to `RUN` on `RUN_INTEGRATION_TESTS` to define the purpose better. Strtobool and read str env variables correctly to represent proper booleans.
+- `SKIP` updated to `RUN` on `RUN_INTEGRATION_TESTS` to define the purpose better. Added `is_truthy` helper.
 
 ### Removed